Handle transactions on sessions in testing mode.
If the application runs in testing mode a global session is used. A new
session is created and a transaction is started if the request has a
"_testcase" paramenter with the value "begin". The session will stay open and
will be reused for all upcoming requests. This make is possible to modify data
in the open transaction. If the request has the paramenter "_testcase" with
value "end" than the session will be rolled back.