Improved fix for #19
There was an error on saving checkboxes with multible selected values which
will be saved in a serialid form as string in the database. Now only apply
special handling for listvalue if the old value and the new value is of
instance list. Otherwise stay with the old behaviour.

is a list.