Moved help text in auth form into infotext
As the form elements are used in the login form too the help texts would be
show there too, which is not wanted. Therefor it is not good to have the
helptexts attached to the entity. Helptexts are now written as text
elements next to the form element only in the registration form.