Handle db interaction with in POST request within a nested transaction.
When entering the handle_POST_request_method a nested transaction is startet.
This nested transaction is used for all interactions with the database within
the hand_post_request_method. So if there happens a exception only this
transaction is rolled back. This is usable to handle IntegrityErrors like
Unique-Contraints etc.

Why we do this? Only rollíng back the nested transaction will keep the
"parent" transaction still usable. This is crucial as the session may be
accessed when rendering the page.