Fix/Change behavior of the "ensure_logout" tween
Former implementation cases ISEs in some errors. The reason is not clear yet
but seems to be related with the accessof the request.user element within the
tween. When doing this the request.user seems to get detached from the session
and lazy evaluation fails. This was at least true for the plorma application.