Added `strict` flag for the get_value method on base items.
You can set this flag to `False` to suppress error logging in case a attribute
can not be fetched. This flag was added as result of a steady discussion if
logging is correct in cases were we expect the fetching to fail.

Often this happens in overviews where items of related items should be shown:
foo.bar.baz. In some situations it is OK that there is no relation bar but
fetching baz will fail and the error will be logged. In this situations you
can disable the error logging.

The flag defaults to True which means errors will be logged. Often these
errors points to flaws in the datamodel and can and should be handled in other
ways then simply suppressing the error.