Improved UI dialog on forgot password workflow
The dialog in the forgot password workflow now does not contain any
information about the fact if a user with the given username could be found or
not. It just says that if there has been a user the user will have received an
email.

This change is a improvement in security as messages of the system must not
provide security relevant informations.