Changed behavior of the redirect on form submission.
Changed the behaviour as followed:

1. In case the user opens a item (in listingrenderer) and click on "Save" the
user will be redirected back to the item.
2. In case the user opens a item (in listingrenderer) and click on "Save and
proceed" the user will stay in the current form on the current page. When the
user now clicks on "Save" the second time the user will stay in the form. If
the user clicks on "Save and procceed" the user will be redirected on the next
formular page (same behavior as before)

So the meaning of "Save and Procceed" has changed when having a backlink in
the way that the userwants to procced in the current form.