Performnace: Optimze loading BaseLists
Add optional user parameter to the BaseList. If provided than a optimzed
SQL-query will be done which only loads items into the list where the uid of
the entry or group matches the given user.

This is used in the get_item_method to optimze the loading time of large
overviews.