Only include ace.js (big) on a very few request.
Currently we need the ace library only for the formeditor. So we only add it
to the header if the formeditor is called. ("/forms/create|update").

Unfortunately we need to to this in ringo instead of formbar. This is because formbar is not
aware of building the correct filepaths to include the scripts on itsself.
Otherwise we would bind the needed JS code to the formeditor field in formbar.