Be more tolerant on values in "data" attribute
If the serializer found a value in the "data" attribute it assumed that this
is the data attribute from the blobform which stored as json string in it. At
least the ringo_file extension has a "data" attribute too. In contrast to the
data attribute it does not store any JSON, so we add a special handling for
normal values here.