Removed "user" parameter from baselist
This attribute was introduced to do some permormance optimization on the
permission checks to only return items which are owned by the user or
the user was member of. Later we noticed that filtering on this level we
will break the option to write custom permission checks in the
_get_permission class method and that this optimization doesn't bring
too much performance. So will removed again to keep the code clean.