Implemented opimized item loading for list views.
Added "load_items" methods into the the base list views to do opimized loading
of items on paginated overview pages. Optimzed means to do as much sorting,
searching as possible on the database an only fetch the relevant slice of data
from the database to reduce the loading on the application.

As it is not possible yet to support all options from the usual way to do
sorting and searching on application side the optimzed loading can only be
applied on a subset of search and sort usecases. In all other cases we will
fall back into the old way of loading.

If opimized loading is active it brings a significant performance boost.