Basically support locale date formats in advanced search.
In default search date objects and other pythonic values are prettified before
appplying the search to support local dateformats e.g. In case the user is
using one of the supported operators, the pythonic value are only converted to
strings. This ensures that expressions like "< 2017-01-01" can be done
because of a simple string comparison.