Add history an NewRequest event again.
Reverted previous change. Basically it is more sensefull to only add entries
to the history in case the server invokes a new Response with 200 status code
but unfortunately at this time we can not store the session so the histroy
stays empty.

For this reason we must add entries to the historey an every new request. This
means that internally the last recent entry in the history is the current url.
This fact is now also handled in the pop and last method of the history.