Enhance redirect handling for update actions.
In case of a update action we need to check if the user is still allowed to
open the update page. This might change due to changes in the status modell
were the state changes and the current user isn't allowed anymore to update
the item and can only read it.