Check if readmode is configured on url generation
There are two major places in ringo where urls are generated. Both methods has
been extended to check if the application is configured to open items in
readmode on default. In this case no further checks are done to check if the
user is allowed to open an item in update mode.