Fix and document handling of relations in (recursive) imports.
The relations build up in one step should not be deleted in another step.
autoflush is well needed in case an item appears multiple times in a
nested JSON file. As the relations are not destroyed anymore, SQLAlchemy
also correctly takes care of the foreign keys so that no integrity errors
should occur on INSERTs anymore.